New Metadata Struct
Finally I came up with the best design for the metadata so far. They are
a little different from the one we discussed before.

typedef struct{
unsigned int size:31;
unsigned int alloc_flag:1;
unsigned int prev_index:31;
unsigned int last_flag:1;
}Metadata

In the previous design, the length of prev_index and next_index are the
bottle neck for this implementation. Even if size can hold up to 31
bits, prev and next references can only hold 16 bits.
This new design solves the problem. It stores prev index and a flag to
tell if this block is the last one.
Two utility functions are used to return the prev and next pointer.
Prev pointer is directly get from the index.
Next pointer is calculated with the current pointer and size.
last_flag is very useful for traversing the list.

I also wrote two utility functions to get the pointers of prev and next
block. They'll be included in mymalloc.c later.